requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==19.3.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
Django==3.2.6
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.4
Flask==2.0.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pyflakes==2.2.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pytz==2021.1
screen==1.0.1
sqlparse==0.3.1
urllib3==1.26.7
Werkzeug==2.0.2
xlwt==1.3.0
